LP2003106 ASN supports repeated container codes per file
If an ASN EDI file contains multiple EDI messages, which may contain
multiple occurrences of a shipment notification container code, append
the subsequent line items to the existing shipment notification, so long
as they are in the same EDI file.

Previously, the code would result in an error regarding duplicate
container codes, which must be unique per provider in the database.

Signed-off-by: Bill Erickson <berickxx@gmail.com>